Moved extractArray to utils, added imports of tensor models to init
